requirements.txt
google.generativeai
langchain-community
gradio